refactor(Car, CarName): 자동차 이름 원시값을 객체로 포장
